+++ Compiler_V3/Parser/transformer.py
@@ -1,15 +1,6 @@
-from typing import List, Dict, Any
 from lark import Transformer, Token
 
-
-class ASTNode:
-    def __init__(self, tag: str, children: List['ASTNode'] = None, attributes: Dict[str, Any] = None):
-        self.tag = tag  # Stores the tag name (e.g., 'row', 'box')
-        self.children = children if children is not None else []
-        self.attributes = attributes if attributes is not None else {}
-
-    def __repr__(self):
-        return f"ASTNode(tag={self.tag}, attributes={self.attributes}, children={len(self.children)})"
+from Compiler_V3.models import ASTNode
 
 
 class DSLTransformer(Transformer):
